% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addalpha.R
\name{addalpha}
\alias{addalpha}
\title{Add transparency (alpha) to a vector of colours}
\usage{
addalpha(x, alpha = 1)
}
\arguments{
\item{x}{A vector of colours (names, hex, etc)}

\item{alpha}{The required alpha. 0 is fully transparent, 1 is fully opaque.}
}
\value{
A vector of colours of the same length as x, with transparency added.
}
\description{
This function accepts a vector of colours and adds a given alpha to make them 
transparent.
}
\examples{
addalpha(c("red", "green", "blue"), 0.3)
addalpha(rainbow(8), 0.6)

}
