% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurve.R
\name{epicurve}
\alias{epicurve}
\title{Draw a classic style epidemic curve (with rectangles)}
\usage{
epicurve(x, series = NA, col = "green", xlim = range(x, na.rm = TRUE) +
  xmargin * c(-1, 1), xlab = NA, ylab = NA, xaxt = "s", yaxt = "s",
  box = FALSE, xmargin = 2)
}
\arguments{
\item{x}{Data vector of integers or 'Date' objects}

\item{series}{Factor of equal length as x, used to group cases and color them separately}

\item{col}{Color, or vector of colors of length equal to the levels of \code{series}.}

\item{xlim}{Limits for the x axis. Defaults to the range of \code{x}, 
plus/minus \code{xmargin}}

\item{xlab}{Title for x axis}

\item{ylab}{Title for y axis}

\item{xaxt}{x axis type. Specifying \code{xaxt="n"} suppresses plotting of the x axis.
You can then draw it explicitly with \code{link{axis}}.}

\item{yaxt}{y axis type. Specifying \code{yaxt="n"} suppresses plotting of the y axis.
You can then draw it explicitly with \code{link{axis}}.}

\item{box}{If FALSE, rectangles are drawn with the appropriate aspect ratio to fit 
the range of both axes. If TRUE, the function adjusts the length of the y axis
so that squares are drawn}

\item{xmargin}{When drawing the x axis and the limits are not explicitly set, extend
the range of x by this amount and use that as the limits of the x axis.}
}
\value{
Nothing. The function plots the epidemic curve on the current graphics device.
}
\description{
epikml() takes a set of coordinates and other associated info as input, and creates
a KML (Keyhole Markup Language) file that can be opened with Google Earth or other 
similar programs. It's original intention was to plot disease cases, but can find wider 
use as well.
}
\details{
Scales are NOT drawn for either axis. Use \code{\link{axis}} to draw them
   explicitly to your liking.
}
\examples{
# Create some dummy data
sampdates <- seq(as.Date("2016-07-15"), as.Date("2016-09-15"), 1)
x <- sample(sampdates, 120, rep=TRUE)
gender <- sample(c("Male","Female"), 120, rep=TRUE)

# Draw the epidemic curve
epicurve(x, gender, c("skyblue","pink"))

}
