% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI.R
\name{getCI}
\alias{getCI}
\title{Get point estimates and 95% Confidence Intervals from a regression model}
\usage{
getCI(m, intercept = FALSE, prob = 0.95)
}
\arguments{
\item{m}{A regression model}

\item{intercept}{Include the intercept in the result?}

\item{prob}{Default 0.95 for a 95% Confidence Interval}
}
\value{
A data.frame with three columns: \code{est}, \code{lo}, \code{hi} providing 
   the point estimate and lower/upper limits of the 95% Confidence Interval for each 
   model parameter.

   The function supports objects of class 'lm' and 'glm', 'lmerMod' and 'glmerMod' 
   (mixed-effects models from package \code{lme4}, for which only the 
   fixed-effects part is considered) and 'coxph' (from package \code{survival})
}
\description{
This function generates point estimates and 95% Confidence Intervals for 
the parameters of a regression model.
}
\examples{
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
mylogit <- glm(admit ~ gre + gpa + rank, data = mydata, family = "binomial")
getCI(mylogit)

}
