% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{Calculate a simple moving average}
\usage{
ma(x, d, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{d}{Window size}

\item{na.rm}{Use a shorter window size for the first values? See "Return Value".}
}
\value{
A vector of the same length as x. If \code{na.rm=FALSE} (the default),
   the first \code{d-1} places of the return vector are \code{NA}. If \code{na.rm=TRUE},
   a moving average is calculated for these with a shorter window size.
}
\description{
This function calculates a simple moving average
}
\examples{
ma(1:6, 3)
# Returns NA NA 2 3 4 5

ma(1:6, 3, na.rm=TRUE)
# Returns 1.0 1.5 2.0 3.0 4.0 5.0

}
