% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Multiple argument version of gsub()}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{A character vector (possibly of length >1) of patterns to be matched, 
see \link{\code{gsub}}}

\item{replacement}{A character vector (of same length as \code{pattern}) with replacement
for each matched pattern. See \link{\code{gsub}}}

\item{x}{a character vector where matches are sought, see \link{\code{gsub}}}

\item{...}{further arguments passed to \link{\code{gsub}} (such as \code{fixed}, for example)}
}
\value{
The target character vector \code{x} with each matching pattern replaced in sequence.
}
\description{
This functions does pattern matching and replacement on a character vector 
just like gsub(), but accepts multiple pattern arguments that are replaced 
sequentially on the target character vector.
}
