% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameMe.R
\name{nameMe}
\alias{nameMe}
\title{Name a vector on-the-fly}
\usage{
nameMe(x, names)
}
\arguments{
\item{x}{A vector (of any type)}

\item{names}{A vector of names, with the same length as \code{x}}
}
\value{
Just \code{x} with the \code{names} assigned.
}
\description{
This function gets a vector and a name(s), and returns the vector with the names assigned.
}
\details{
This function is useful when one uses c() to concatenate complex arguments, but 
does not want to assign these to temporary objects in order to name them.
}
\examples{
c(one=10, nameMe(c(3.14, 5), c("two", "three")))

}
