% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColors.R
\name{plotColors}
\alias{plotColors}
\alias{plotColours}
\title{Plot colors from R's color palette}
\usage{
plotColors(x)

plotColours(x)
}
\arguments{
\item{x}{One or more color names (as a character vector)}
}
\value{
Nothing
}
\description{
This function takes a color name -or more- as argument (e.g. "green") and 
displays all the variants of this color that R knows about. This way one 
can choose the most aesthetically pleasing colors to use for his/her plot.
}
\details{
The function matches x to the color names that \code{\link{colors}} 
   returns, and then plots nice rectangles each filled with a matching color 
   and labelled with the respective color name.
}
\examples{
plotColors("sea")
plotColors(c("green","yellow"))

}
