% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoweekStart.R
\name{seq_isoweek}
\alias{seq_isoweek}
\title{Generate regular sequences of ISO weeks}
\usage{
seq_isoweek(from, to = NULL)
}
\arguments{
\item{from}{Numeric vector of length 1 (first week of sequence), or of length 2
(week range) if \code{to} is \code{NULL}}

\item{to}{Numeric vector of length 1 (last week of sequence), or \code{NULL} (in which
case \code{from} must be a range of weeks).}
}
\value{
A numeric vector of ISO weeks covering the provided range.
}
\description{
Take two ISO week numbers (of the form YYYYWW) and returns a regular sequence of 
ISO weeks covering the entire range.
}
\examples{
seq_isoweek(c(202001, 202301))
seq_isoweek(201840, 201920)  # Typical influenza surveillance period

}
