% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{vecshift}
\alias{vecshift}
\title{Create lagged or lead variables}
\usage{
vecshift(x, d = 0)
}
\arguments{
\item{x}{A vector (of any type)}

\item{d}{How many places to shift. A positive integer means a shift to the left,
a negative integer means a shift to the right.}
}
\value{
A vector of the same length as x, shifted \code{d} places to the left (if 
   \code{d} positive) or right (if \code{d} negative). The rightmost or leftmost
   \code{abs(d)} place respectively are filled with \code{NA}.
}
\description{
This function replicates the functionality of the Stata subscript syntax (e.g. [_n-1])
to positionally shift a vector and create lagged or lead variables.
}
\details{
This function calculates the week number according to ISO 8601. Note that
   dates near the start or end of a given year may belong to the previous or next
   year respectively, thus the year needs to be calculated too.
}
\examples{
vecshift(1:10, 3)

}
