% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoef.R
\name{getCoef}
\alias{getCoef}
\title{Get estimates for a parameter combination from a regression model}
\usage{
getCoef(m, par)
}
\arguments{
\item{m}{A regression model (of class 'lm','glm','lmerMod','glmerMod' or 'coxph')}

\item{par}{A matrix (or data.frame) specifying the parameter combinations.}
}
\value{
A data.frame with three columns: \code{est}, \code{lo}, \code{hi} providing 
   the point estimate and lower/upper limits of the 95% Confidence Interval, for each 
   parameter combination provided in \code{par}. Thus the return value has the same 
   number of rows as \code{par}.

   The function supports objects of class 'lm' and 'glm', 'lmerMod' and 'glmerMod' 
   (mixed-effects models from package \code{lme4}, for which only the 
   fixed-effects part is considered) and 'coxph' (from package \code{survival})
}
\description{
This function generates point estimates and 95% Confidence Intervals from a 
regression model, for any combination of model parameters.
}
\examples{
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
mylogit <- glm(admit ~ gre + gpa + rank, data = mydata, family = "binomial")
pars <- data.frame(gpa=c(2,2,3), rank=c(3,4,4))
getCoef(mylogit, pars)

}
