% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piformat.R
\name{piformat}
\alias{piformat}
\title{Pretty-printer for p-values}
\usage{
piformat(x, html = FALSE, tb = TRUE)
}
\arguments{
\item{x}{One or more color names (as a character vector)}

\item{html}{Escape the < character with the &lt; character entity}

\item{tb}{Omit the "p=" or "p<" part, e.g. if outputting to a table}
}
\value{
A character vector of equal length to x, with formatted p-values
}
\description{
This function formats p-values in the most usual way, i.e. with two decimals if 
p>=0.05, with three decimals if p>=0.001 and p<0.05, and as "p<0.001" if p<0.001.
}
\examples{
piformat(c(0.0001,0.001,0.01,0.1,0.5))

}
